refactor option parsing
