+++ src/Cauterize/GHC7/Options.hs
@@ -1,7 +1,10 @@
 module Cauterize.GHC7.Options
   ( CautGHC7Opts(..)
   , modulePathAsList
+  , options
   ) where
+
+import Options.Applicative
 
 data CautGHC7Opts = CautGHC7Opts
   { specFile :: FilePath
@@ -18,4 +21,31 @@ modulePathAsList opts = splitWith (== '.') <$> modulePath opts
       where (first, rest) = break cond xs
   safeTail [] = []
   safeTail (_:ys) = ys
+
+optParser :: Parser CautGHC7Opts
+optParser = CautGHC7Opts
+  <$> argument str
+    ( metavar "SPEC"
+   <> help "Cauterize specification"
+    )
+  <*> argument str
+    ( metavar "OUTDIR"
+   <> help "Output directory"
+    )
+  <*> option (Just <$> str)
+    ( long "module-path"
+   <> metavar "Module.Path"
+   <> help "override default haskell module path"
+   <> value Nothing
+    )
+
+options :: ParserInfo (Maybe CautGHC7Opts)
+options = info (helper <*> o)
+   ( fullDesc
+  <> progDesc "Translate a Cauterize specification into a Haskell implementation"
+   )
+  where
+  o = flag' Nothing (long "version" <> hidden)
+   <|> (Just <$> optParser)
+
