+++ src/Main.hs
@@ -4,36 +4,20 @@ module Main
 
 import Cauterize.GHC7.Generate
 import Cauterize.GHC7.Options
+import Cauterize.GHC7.Version
 
-import Options.Applicative
+import Options.Applicative (execParser)
 
 main :: IO ()
 main = runWithOptions caut2hs
 
 runWithOptions :: (CautGHC7Opts -> IO ()) -> IO ()
-runWithOptions fn = execParser options >>= fn
+runWithOptions fn = do
+  mopts <- execParser options
+  case mopts of
+    Just opts -> fn opts
+    Nothing -> do
+      putStr versionString
+      putStrLn "dependencies:"
+      putStr dependencyString
 
-optParser :: Parser CautGHC7Opts
-optParser = CautGHC7Opts
-  <$> strOption
-    ( long "spec"
-   <> metavar "SPEC_PATH"
-   <> help "Cauterize specification file"
-    )
-  <*> strOption
-    ( long "output"
-   <> metavar "OUTPUT_PATH"
-   <> help "Directory to save output files"
-    )
-  <*> option (Just <$> str)
-    ( long "module-path"
-   <> metavar "Module.Path"
-   <> help "override default haskell module path"
-   <> value Nothing
-    )
-
-options :: ParserInfo CautGHC7Opts
-options = info (optParser <**> helper)
-            ( fullDesc
-           <> progDesc "Translate a Cauterize specification and meta file into a Haskell library."
-            )
